+++ test/features/steps/auto_vertical.py
@@ -3,11 +3,12 @@ from textwrap import dedent
 from behave import then, when
 
 import wrappers
+from utils import parse_cli_args_to_dict
 
 
 @when('we run dbcli with {arg}')
 def step_run_cli_with_arg(context, arg):
-    wrappers.run_cli(context, run_args=arg.split('='))
+    wrappers.run_cli(context, run_args=parse_cli_args_to_dict(arg))
 
 
 @when('we execute a small query')