+++ changelog.md
@@ -1,9 +1,10 @@
-TBD
-=======
+TODO
+====
 
 Bug Fixes:
 ----------
 * Allow `FileNotFound` exception for SSH config files.
+* Fix startup error on MySQL < 5.0.22
 * Check error code rather than message for Access Denied error
 
 Features:
@@ -13,6 +14,7 @@ Features:
 Internal:
 ---------
 * Remove unused function is_open_quote()
+* Test various host-port combinations in command line arguments
 
 
 1.23.2